Moved UI from internal to pkg (will be used by dependencies)
