+++ cmd/main.go
@@ -6,12 +6,12 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/eko/monday/internal/ui"
 	"github.com/eko/monday/pkg/config"
 	"github.com/eko/monday/pkg/forwarder"
 	"github.com/eko/monday/pkg/hostfile"
 	"github.com/eko/monday/pkg/proxy"
 	"github.com/eko/monday/pkg/runner"
+	"github.com/eko/monday/pkg/ui"
 	"github.com/eko/monday/pkg/watcher"
 	"github.com/manifoldco/promptui"
 	"github.com/spf13/cobra"
