+++ pkg/forwarder/forwarder.go
@@ -8,10 +8,10 @@ import (
 	"time"
 
 	"github.com/eko/monday/pkg/config"
-	"github.com/eko/monday/internal/ui"
 	"github.com/eko/monday/pkg/forwarder/kubernetes"
 	"github.com/eko/monday/pkg/forwarder/ssh"
 	"github.com/eko/monday/pkg/proxy"
+	"github.com/eko/monday/pkg/ui"
 )
 
 // ForwarderInterface represents all kinds of forwarders (Kubernetes, others...)
