+++ pkg/runner/runner.go
@@ -9,8 +9,8 @@ import (
 	"syscall"
 
 	"github.com/eko/monday/pkg/config"
-	"github.com/eko/monday/internal/ui"
 	"github.com/eko/monday/pkg/proxy"
+	"github.com/eko/monday/pkg/ui"
 )
 
 var (
