+++ pkg/ui/view_test.go
